Refactoring alignment rules:
  1)  Created utils file specifically for alignment rules
  2)  Refactored rules to move identical code into utils file
